Renamed PieChartTest to PieChartView
